+++ Web/Scotty/Trans/Lazy.hs
@@ -17,12 +17,14 @@ raise :: (MonadIO m) =>
          T.Text -- ^ Error text
       -> ActionT m a
 raise  = Base.raise . T.toStrict
+{-# DEPRECATED raise "Throw an exception instead" #-}
 
 -- | Throw a 'StatusError' exception that has an associated HTTP error code and can be caught with 'rescue'.
 --
 -- Uncaught exceptions turn into HTTP responses corresponding to the given status.
 raiseStatus :: Monad m => Status -> T.Text -> ActionT m a
 raiseStatus s = Base.raiseStatus s . T.toStrict
+{-# DEPRECATED raiseStatus "Use status, text, and finish instead" #-}
 
 -- | Redirect to given URL. Like throwing an uncatchable exception. Any code after the call to redirect
 -- will not be run.